Restructure Windows monkey patch for Open3. #4943
