+++ lib/ruby/stdlib/open3.rb
@@ -29,14 +29,7 @@
 # - Open3.pipeline : run a pipeline and wait for its completion
 #
 
-# Because spawn does not yet work on Windows, we fall back on the older open3 there.
-real_open3 = true
-if respond_to?(:org) && org.jruby.platform.Platform::IS_WINDOWS
-  require 'jruby/open3_windows'
-  real_open3 = false
-end
-
-real_open3 && module Open3
+module Open3
 
   # Open stdin, stdout, and stderr streams and start external executable.
   # In addition, a thread to wait for the started process is created.
@@ -673,4 +666,9 @@ real_open3 && module Open3
     private :pipeline_run
   end
 
+end
+
+# Because spawn does not yet work on Windows, we fall back on the older open3 there.
+if RUBY_ENGINE == 'jruby' && org.jruby.platform.Platform::IS_WINDOWS
+  require 'jruby/open3_windows'
 end
